pyproject.toml
[project]
name = "engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core web framework
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "pydantic>=2.11.7",
    "python-dotenv>=0.9.9",
    # LangChain and workflow
    "langchain-core>=0.3.75",
    "langchain-ollama>=0.3.7",
    "langgraph>=0.6.7",
    # ML and experiment tracking
    "mlflow>=3.3.2",
    "tensorflow>=2.20.0",
    "tf-keras>=2.20.1",
    "numpy>=2.3.2",
    "opencv-python-headless>=4.11.0.86",
    # Critical transitive dependencies (needed for protobuf compatibility)
    "protobuf>=5.28.0",
    "grpcio>=1.50.0",
    "python-multipart>=0.0.20",
    "opencv-python>=4.11.0.86",
    "langchain>=0.3.27",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "matplotlib>=3.10.6",
    "scikit-image>=0.25.2",
    "kagglehub>=0.3.13",
    "transformers>=4.56.2",
    "requests>=2.32.5",
    "opentelemetry-api==1.22.0",
    "opentelemetry-sdk==1.22.0",
    "opentelemetry-instrumentation==0.43b0",
    "opentelemetry-instrumentation-fastapi==0.43b0",
    "opentelemetry-instrumentation-httpx==0.43b0",
    "opentelemetry-instrumentation-requests==0.43b0",
    "opentelemetry-instrumentation-logging==0.43b0",
    "opentelemetry-instrumentation-system-metrics==0.43b0",
    "opentelemetry-exporter-prometheus==0.43b0",
    "opentelemetry-semantic-conventions==0.43b0",
    "opentelemetry-instrumentation-redis==0.43b0",
    "opentelemetry-exporter-otlp>=1.11.1",
]
